+++ examples/apps/react-todomvc-rtk/src/store/actions.js
@@ -1,0 +1,4 @@
+import { createAction } from "@reduxjs/toolkit"
+
+export const formSubmit = createAction("formSubmit")
+export const clearClick = createAction("clearClick")