+++ examples/apps/web-select/CHANGELOG.md
@@ -1,5 +1,11 @@
 # web-select
 
+## 0.2.27
+
+### Patch Changes
+
+- @inglorious/web@4.2.4
+
 ## 0.2.26
 
 ### Patch Changes