+++ packages/logo/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inglorious/logo
 
+## 2.0.33
+
+### Patch Changes
+
+- @inglorious/web@4.2.4
+
 ## 2.0.32
 
 ### Patch Changes