+++ packages/react-store/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @inglorious/react-store
 
+## 16.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/store@9.5.3
+
 ## 16.0.2
 
 ### Patch Changes