+++ packages/renderer-2d/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inglorious/renderer-2d
 
+## 19.0.3
+
+### Patch Changes
+
+- @inglorious/engine@20.0.3
+
 ## 19.0.2
 
 ### Patch Changes